requirements.txt
passlib==1.7.4
sqlalchemy==2.0.30
pyjwt==1.7.1
fastapi==0.111.0
alembic==1.13.1
bcrypt==4.1.3
python-environ==0.4.54